+++ cookbooks/cobbler/metadata.rb
@@ -1,6 +1,6 @@
-maintainer       "Webtrends"
-maintainer_email "optimize.operations-team@Webtrends.com"
+maintainer       "Webtrends Inc."
+maintainer_email "hostedops@webtrends.com"
 license          "Apache 2.0"
 description      "Installs/Configures cobbler"
-long_description IO.read(File.join(File.dirname(__FILE__), 'README.rdoc'))
+long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version          "0.0.1"