+++ cookbooks/kafka/attributes/default.rb
@@ -34,4 +34,3 @@ default[:kafka][:log_flush_scheduler_time_interval] = 1000
 default[:kafka][:log_retention_hours] = 168
 
 
-