+++ cookbooks/ms_dotnet4/metadata.rb
@@ -5,5 +5,5 @@ license          "All rights reserved"
 description      "Installs/Configures Microsoft .NET 4.0"
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version          "1.0.1"
-supports	 "windows"
+supports         "windows"
 depends          "windows"