+++ cookbooks/wt_streamingconfigservice/attributes/default.rb
@@ -1,2 +1,11 @@
+#
+# Cookbook Name:: wt_streamingconfigservice
+# Attributes:: default
+#
+# Copyright 2012, Webtrends
+#
+# All rights reserved - Do Not Redistribute
+#
+
 default['wt_streamingconfigservice']['user'] = "webtrends"
 default['wt_streamingconfigservice']['group'] = "webtrends"