+++ cookbooks/zookeeper/attributes/default.rb
@@ -4,8 +4,6 @@
 #
 # Copyright 2012, Webtrends Inc.
 #
-# All rights reserved - Do Not Redistribute
-#
 
 default[:zookeeper][:version] = "3.3.6"
 default[:zookeeper][:download_url] = "http://mirror.uoregon.edu/apache/zookeeper/zookeeper-#{node[:zookeeper][:version]}/zookeeper-#{node[:zookeeper][:version]}.tar.gz"